requirements.txt
discord.py>=2.1.0
tensorflow>=2.8.0
numpy>=1.22.0
pillow>=9.0.0
matplotlib>=3.5.0